router/requirements.txt
# AlphaOS Router Bot Dependencies

aiogram>=3.0.0
aiohttp>=3.8.0
PyYAML>=6.0
